first commit o

+++ src/FindMax.java
@@ -5,17 +5,18 @@ public class FindMax {
         Scanner scanner = new Scanner(System.in);
 
         System.out.println("Enter val1 : ");
-        int val1 = scanner.nextInt();
+        double val1 = scanner.nextDouble();
 
         System.out.println("Enter val2 : ");
-        int val2 = scanner.nextInt();
+        double val2 = scanner.nextDouble();
 
         System.out.println("Enter val3 : ");
-        int val3 = scanner.nextInt();
+        double val3 = scanner.nextDouble();
 
         if(val1 > val2 && val1 > val3) {
             System.out.println(val1 + " is maximum");
         }
+
             else if(val2 > val1 && val2 > val3){
                 System.out.println(val2 + " is maximum");
             }
